move streaming functionality to its own file
