+++ setup.py
@@ -10,18 +10,18 @@ DESCRIPTION = (
     "processing")
 
 SCRIPTS = [
-    'rosetta/cmdutils/' + name for name in 
+    'rosetta/cmdutils/' + name for name in
     ['concat_csv.py', 'cut.py', 'join_csv.py',
     'row_filter.py', 'split.py', 'subsample.py',
     'files_to_vw.py', 'filter_sfile.py', 'groupby_reduce.py']]
 
 PACKAGES =  ['rosetta'] + [
-    'rosetta.' + name for name in 
+    'rosetta.' + name for name in
     ['cmdutils', 'modeling', 'parallel', 'text', 'workflow']]
 
 
 #PY_MODULES =  ['rosetta'] + [
-#    'rosetta/' + name for name in 
+#    'rosetta/' + name for name in
 #    ['cmdutils', 'modeling', 'parallel', 'text', 'workflow']]
